refactor(monitoring): extract sub-components from AdvancedPerformanceDashboard
- Created centralized types file (advanced-performance-dashboard.types.ts)
- Extracted PerformanceOverviewTab component (140 lines)
- Extracted AIOptimizationTab component (100 lines)
- Extracted PredictiveAnalyticsTab component (125 lines)
- Refactored main component to use extracted modules (78% code reduction)
- Replaced hardcoded interval with STANDARD_INTERVALS hook
- Updated monitoring-api.ts imports to use centralized types
- All quality gates passing: Build ✅, Lint ✅, Typecheck ✅, Tests ✅ (42/42, 535/535)
