+++ lib/services/monitoring-api.ts
@@ -2,7 +2,7 @@ import type {
   AdvancedPerformanceMetrics,
   AICacheOptimizationMetrics,
   PredictivePerformanceData,
-} from "@/components/monitoring/advanced-performance-dashboard";
+} from "@/components/monitoring/advanced-performance-dashboard.types";
 
 export interface WebhookQueueStats {
   queue: {
